Include user rolesin

+++ src/main/java/com/toolkit/inventory/Security/Domain/Role.java
@@ -21,7 +21,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Column(name = "role_name")
     private String roleName;
